+++ examples/Common/Events/PaymentCaptured.cs
@@ -10,7 +10,7 @@ public class PaymentCaptured : IEvent
 
     public int Amount { get; init; }
 
-    public DateTime Timestamp { get; init; }
+    public DateTimeOffset Timestamp { get; init; }
 
     public IReadOnlyDictionary<string, string>? Metadata { get; set; }
 }