+++ examples/Common/Events/PaymentDeclined.cs
@@ -10,7 +10,7 @@ public class PaymentDeclined : IEvent
 
     public string? Reason { get; init; }
 
-    public DateTime Timestamp { get; init; } = DateTime.UtcNow;
+    public DateTimeOffset Timestamp { get; init; } = DateTimeOffset.UtcNow;
 
     public IReadOnlyDictionary<string, string>? Metadata { get; set; }
 }