+++ examples/Common/Events/RefundActionAccepted.cs
@@ -8,7 +8,7 @@ public class RefundActionAccepted : IEvent
 {
     public string TransactionId { get; init; } = null!;
 
-    public DateTime Timestamp { get; init; }
+    public DateTimeOffset Timestamp { get; init; }
 
     public IReadOnlyDictionary<string, string>? Metadata { get; set; }
 }