+++ examples/Common/Events/RefundDeclined.cs
@@ -16,7 +16,7 @@ public class RefundDeclined : IEvent
 
     public string? Reason { get; init; }
 
-    public DateTime Timestamp { get; init; }
+    public DateTimeOffset Timestamp { get; init; }
 
     public IReadOnlyDictionary<string, string>? Metadata { get; set; }
 }