+++ examples/Common/Events/RefundRequestedRepeated.cs
@@ -14,7 +14,7 @@ public class RefundRequestedRepeated : IEvent
 
     public int? Amount { get; init; }
 
-    public DateTime Timestamp { get; init; }
+    public DateTimeOffset Timestamp { get; init; }
 
     public IReadOnlyDictionary<string, string>? Metadata { get; set; }
 }