+++ src/Marty.Net.Contracts/IEventStore.cs
@@ -174,7 +174,7 @@ public interface IEventStore
     /// <returns>A list of tuples containing the events with their associated (optional) metadata.</returns>
     Task<List<IEvent>> ReadStreamUntilTimestamp(
         string streamName,
-        DateTime timestamp,
+        DateTimeOffset timestamp,
         CancellationToken cancellationToken = default
     );
 
@@ -187,7 +187,7 @@ public interface IEventStore
     /// <returns>A list of tuples containing the events with their associated (optional) metadata.</returns>
     Task<List<IEvent>> ReadStreamFromTimestamp(
         string streamName,
-        DateTime timestamp,
+        DateTimeOffset timestamp,
         CancellationToken cancellationToken
     );