+++ tests/Marty.Net.Tests/Events/Orders/v2/OrderEventCancelled.cs
@@ -12,7 +12,7 @@ public class OrderEventCancelled : IEvent
 
     public string? Reason { get; init; }
 
-    public DateTime Timestamp { get; init; } = DateTime.UtcNow;
+    public DateTimeOffset Timestamp { get; init; } = DateTimeOffset.UtcNow;
 
     public IReadOnlyDictionary<string, string>? Metadata { get; set; }
 }