+++ RELEASE_NOTES.md
@@ -1,5 +1,12 @@
 Release Notes
 
+## [1.0.1](https://github.com/dariogriffo/easy-evs/releases/tag/1.0.1)
+
+Use DateTimeOffset instead of DateTime for the event timestamp.
+
+Fixed lifetime for IAggregateStore
+
+
 ## [1.0.0](https://github.com/dariogriffo/easy-evs/releases/tag/1.0.0)
 
 First version with a stable API.